[87] fix clock images
Fixes the clock images. Presumably, the dependency version changes
introduced in a53ea2824e02ff84b8edfc55000f91b5b928e352 made it so that
the onLoad callback didn't fire if the element in question isn't
rendered. Swapping a hidden attribute for an opacity of 0 allows the
element to render even if it isn't ready, and only show once onLoad
fired, fixing the issue.